+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Enhancements:
 * [ORM] Changed Model::deleteEach to use iterator instead of fetching all elements at once (issue #254).
 * [ORM] Fixed fetchIterator to properly use cursor underneath.
 * [DI] Added constructor injection for arguments with types defined (issue #265).
+* [DI] Implemented injection of private fields for parent class.
 * [Utilities] Added equalsIgnoreCase to Functions and FluentFunctions (issue #263).
 * [Utilities] Fixed Clock to support DST changes when adding hours, minutes or seconds.